test(web): fix regression in suggestion setup that caused the suggestion API to error out
The Settings app data entry was of the wrong type, and so the request that tried to load it got an
error instead. Additionally, this code wasn't covered by tests. This corrects both problems: The
request handler requests the correct type for the app data settings, and there is a test that covers
the case.

regressed by #23