test: Re-enable doctest, and mark tests as no_run
One of the benefits of rustdoc is that it can compile your examples so that they stay in sync with your code. However, some of our examples cannot be used as doctests, since they would never exit. The correct way to handle this is with the `no_run` "language" on the code examples, not by disabling rust doc entirely.