Fix the integration tests
The Kinto server was not properly configured
to track changes on non-main buckets. In tests
we created collections and buckets on the fly
and changes to these were not properly tracked.
By creating collections on the fly exclusively
in the 'main' bucket we make sure that any change
is tracked by Kinto and that everything works
as expected