fix(deduped-map): Correctly remove pointer and storage entries when calling DedupedMap::remove
This was likely causing #289, since it causes some storage entries to continue
being stored past when their last pointer was removed, due to the reference
count not being updated correctly.

The result of this is that storage entries would leak if their last pointer was
removed via `::remove` instead of via `::retain`.