Allow disabling deployment
If the latest commit on the PR contains the
[do not deploy] text, the CI tasks to publish
the docker image and trigger deployment will
be skipped.

This means that no deployment will take place.

Please note that only the commit message of the
latest commit will be used as a reference, nor
the ones from the older commits or the PR title.